fix edit bug and split editinvoice file
