+++ frontend/src/components/invoice/InvoiceInformation.tsx
@@ -1,0 +1,174 @@
+import React from "react"
+import { format } from "date-fns"
+import { FormField, FormItem, FormControl, FormLabel, FormMessage } from "@/components/ui/form"
+import { Button } from "@/components/ui/button"
+import { Calendar } from "@/components/ui/calendar"
+import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
+import SectionHeader from "@/components/common/SectionHeader"
+import { CalendarIcon, User, Clock, CalendarDays, FileText } from "lucide-react"
+import { cn } from "@/lib/utils"
+
+interface InvoiceInformationProps {
+  form: any  
+  customers: any[] 
+  paymentTerms: any[]
+  isLoadingCustomers: boolean
+  isLoadingPaymentTerms: boolean
+}
+
+const InvoiceInformation: React.FC<InvoiceInformationProps> = ({
+  form,
+  customers,
+  paymentTerms,
+  isLoadingCustomers,
+  isLoadingPaymentTerms,
+}) => {
+  return (
+    <div>
+      <SectionHeader title="Invoice Information" icon={FileText} />
+      <div className="mt-6 grid grid-cols-1 gap-y-6 gap-x-4 sm:grid-cols-6">
+        {/* Customer */}
+        <FormField
+          control={form.control}
+          name="customer"
+          render={({ field }: any) => (
+            <FormItem className="sm:col-span-3">
+              <FormLabel className="flex items-center gap-1">
+                <User className="h-4 w-4 text-secondary-500" />
+                Customer <span className="text-red-500">*</span>
+              </FormLabel>
+              <FormControl>
+                <select
+                  className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
+                  value={field.value || ""}
+                  onChange={(e) => field.onChange(e.target.value)}
+                  disabled={isLoadingCustomers}
+                >
+                  <option value="">None</option>
+                  {customers.map((customer) => (
+                    <option key={customer.id} value={customer.name}>
+                      {customer.name}
+                    </option>
+                  ))}
+                </select>
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        {/* Invoice Date */}
+        <FormField
+          control={form.control}
+          name="invoiceDate"
+          render={({ field }: any) => (
+            <FormItem className="sm:col-span-3">
+              <FormLabel className="flex items-center gap-1">
+                <CalendarDays className="h-4 w-4 text-secondary-500" />
+                Invoice date <span className="text-red-500">*</span>
+              </FormLabel>
+              <Popover>
+                <PopoverTrigger asChild>
+                  <FormControl>
+                    <Button
+                      variant={"outline"}
+                      className={cn(
+                        "w-full pl-3 text-left font-normal",
+                        !field.value && "text-muted-foreground"
+                      )}
+                    >
+                      {field.value ? (
+                        format(field.value, "dd/MM/yyyy")
+                      ) : (
+                        <span>dd/MM/yyyy</span>
+                      )}
+                      <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+                    </Button>
+                  </FormControl>
+                </PopoverTrigger>
+                <PopoverContent className="w-auto p-0" align="start">
+                  <Calendar
+                    mode="single"
+                    selected={field.value}
+                    onSelect={field.onChange}
+                    initialFocus
+                  />
+                </PopoverContent>
+              </Popover>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        {/* Payment Terms */}
+        <FormField
+          control={form.control}
+          name="paymentTerms"
+          render={({ field }: any) => (
+            <FormItem className="sm:col-span-3">
+              <FormLabel className="flex items-center gap-1">
+                <Clock className="h-4 w-4 text-secondary-500" />
+                Payment terms
+              </FormLabel>
+              <FormControl>
+                <select
+                  className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
+                  value={field.value || ""}
+                  onChange={(e) => field.onChange(e.target.value)}
+                  disabled={isLoadingPaymentTerms}
+                >
+                  <option value="">None</option>
+                  {paymentTerms.map((term) => (
+                    <option key={term.id} value={term.id}>
+                      {term.name}
+                    </option>
+                  ))}
+                </select>
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        {/* Due Date */}
+        <FormField
+          control={form.control}
+          name="dueDate"
+          render={({ field }: any) => (
+            <FormItem className="sm:col-span-3">
+              <FormLabel className="flex items-center gap-1">
+                <CalendarDays className="h-4 w-4 text-secondary-500" />
+                Due date
+              </FormLabel>
+              <Popover>
+                <PopoverTrigger asChild>
+                  <FormControl>
+                    <Button
+                      variant={"outline"}
+                      className="w-full pl-3 text-left font-normal"
+                    >
+                      {field.value ? (
+                        format(field.value, "dd/MM/yyyy")
+                      ) : (
+                        <span>dd/MM/yyyy</span>
+                      )}
+                      <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+                    </Button>
+                  </FormControl>
+                </PopoverTrigger>
+                <PopoverContent className="w-auto p-0" align="start">
+                  <Calendar
+                    mode="single"
+                    selected={field.value}
+                    onSelect={field.onChange}
+                    initialFocus
+                  />
+                </PopoverContent>
+              </Popover>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+      </div>
+    </div>
+  )
+}
+
+export default InvoiceInformation
